requirements.txt
asgiref==3.6.0
Django==4.1.6
django-bootstrap-datepicker-plus==5.0.3
django-humanize==0.1.2
django-widget-tweaks==1.4.12
humanize==4.5.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pydantic==1.10.4
sqlparse==0.4.3
typing-extensions==4.4.0
tzdata==2022.7
